pyproject.toml
[tool.poetry]
name = "nuregi"
version = "0.3.0"
description = "A basic Python-based API client library for registrar.nu.edu.kz"
license = "GPL-3.0-only"
authors = ["aldan <gitaldan@gmail.com>"]
readme = "README.md"
homepage = "https://github.com/aldan/nuregi"
repository = "https://github.com/aldan/nuregi"
classifiers = [
        "Development Status :: 4 - Beta",
        "Intended Audience :: Developers",
        "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
pandas = "^2.1.3"
tabula-py = "^2.9.0"
jpype1 = "^1.4.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
pylint = "^3.0.2"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/aldan/nuregi/issues"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
